"se corrigio el apartado de editar producto, concretamente no aparecia el input llamado Stock, se modifico la entidad, el controller, el dto y el html de productoEditar"

+++ src/main/java/com/kerubinesapp/mateogdev/dto/ProductoDto.java
@@ -14,6 +14,6 @@ public class ProductoDto {
     private String color;
     private String descripcion;
     private Double costo;
-    private Double stock;
+    private Long stock;
 
 }
